Refactor notes handling into dedicated service
