+++ app.py
@@ -48,8 +48,7 @@ from database import (
 from retrorecon import (
     progress as progress_mod,
     saved_tags as saved_tags_mod,
-    notes_utils,
-    text_notes_utils,
+    notes_service,
     search_utils,
     screenshot_service,
     sitezip_utils,
@@ -153,42 +152,6 @@ def save_saved_tags(tags: List[str]) -> None:
     saved_tags_mod.save_tags(SAVED_TAGS_FILE, tags)
 
 
-def get_notes(url_id: int) -> List[sqlite3.Row]:
-    return notes_utils.get_notes(url_id)
-
-
-def add_note(url_id: int, content: str) -> int:
-    return notes_utils.add_note(url_id, content)
-
-
-def update_note(note_id: int, content: str) -> None:
-    notes_utils.update_note(note_id, content)
-
-
-def delete_note_entry(note_id: int) -> None:
-    notes_utils.delete_note_entry(note_id)
-
-
-def delete_all_notes(url_id: int) -> None:
-    notes_utils.delete_all_notes(url_id)
-
-
-def export_notes_data() -> List[Dict[str, Any]]:
-    return notes_utils.export_notes_data()
-
-
-def get_text_notes() -> List[sqlite3.Row]:
-    return text_notes_utils.get_text_notes()
-
-
-def add_text_note(content: str) -> int:
-    return text_notes_utils.add_text_note(content)
-
-def update_text_note(note_id: int, content: str) -> None:
-    text_notes_utils.update_text_note(note_id, content)
-
-def delete_text_note_entry(note_id: int) -> None:
-    text_notes_utils.delete_text_note(note_id)
 
 def export_url_data(ids: Optional[List[int]] = None, query: str = '') -> List[Dict[str, Any]]:
     """Return URL records filtered by ids or search query."""
